Refactor project to MVC structure and update Square.java for JDK 19
