+++ src/Model/Clock.java
@@ -1,3 +1,4 @@
+package Model;
 
 public class Clock {
     private int hh;
