+++ src/Model/piece/Rook.java
@@ -1,11 +1,13 @@
+package Model.piece;
 
+import Model.Board;
 
 import java.util.LinkedList;
 import java.util.List;
 
-public class Queen extends Piece {
+public class Rook extends Piece {
 
-    public Queen(int color, Square initSq, String img_file) {
+    public Rook(int color, Square initSq, String img_file) {
         super(color, initSq, img_file);
     }
 
@@ -27,11 +29,7 @@ public class Queen extends Piece {
             if (i != x) legalMoves.add(board[y][i]);
         }
         
-        List<Square> bMoves = getDiagonalOccupations(board, x, y);
-        
-        legalMoves.addAll(bMoves);
-        
         return legalMoves;
     }
-    
+
 }
