+++ src/Piece.java
@@ -1,5 +1,7 @@
 
 
+import Model.Board;
+
 import java.awt.Graphics;
 import java.awt.Image;
 import java.awt.image.BufferedImage;
