+++ src/Square.java
@@ -1,98 +1,100 @@
+package view;
 
-
-import java.awt.Color;
-import java.awt.Graphics;
+import model.Board;
+import model.Piece;
 
 import javax.swing.*;
+import java.awt.*;
 
-@SuppressWarnings("serial")
 public class Square extends JComponent {
-    private Board b;
-    
+
+    private final Board board;
     private final int color;
+    private final int xNum;
+    private final int yNum;
+
     private Piece occupyingPiece;
-    private boolean dispPiece;
-    
-    private int xNum;
-    private int yNum;
-    
-    public Square(Board b, int c, int xNum, int yNum) {
-        
-        this.b = b;
-        this.color = c;
-        this.dispPiece = true;
+    private boolean displayPiece;
+
+    public Square(Board board, int color, int xNum, int yNum) {
+        this.board = board;
+        this.color = color;
+        this.displayPiece = true;
         this.xNum = xNum;
         this.yNum = yNum;
-        
-        
-        this.setBorder(BorderFactory.createEmptyBorder());
+
+        setBorder(BorderFactory.createEmptyBorder());
     }
-    
-    public int getColor() {
-        return this.color;
+
+    public int getColorCode() {
+        return color;
     }
-    
+
     public Piece getOccupyingPiece() {
         return occupyingPiece;
     }
-    
+
     public boolean isOccupied() {
-        return (this.occupyingPiece != null);
+        return occupyingPiece != null;
     }
-    
+
     public int getXNum() {
-        return this.xNum;
+        return xNum;
     }
-    
+
     public int getYNum() {
-        return this.yNum;
+        return yNum;
     }
-    
-    public void setDisplay(boolean v) {
-        this.dispPiece = v;
+
+    public void setDisplayPiece(boolean displayPiece) {
+        this.displayPiece = displayPiece;
     }
-    
-    public void put(Piece p) {
-        this.occupyingPiece = p;
-        p.setPosition(this);
+
+    public void put(Piece piece) {
+        this.occupyingPiece = piece;
+        piece.setPosition(this);
     }
-    
+
     public Piece removePiece() {
-        Piece p = this.occupyingPiece;
-        this.occupyingPiece = null;
-        return p;
+        Piece removed = occupyingPiece;
+        occupyingPiece = null;
+        return removed;
     }
-    
-    public void capture(Piece p) {
-        Piece k = getOccupyingPiece();
-        if (k.getColor() == 0) b.Bpieces.remove(k);
-        if (k.getColor() == 1) b.Wpieces.remove(k);
-        this.occupyingPiece = p;
-    }
-    
-    public void paintComponent(Graphics g) {
-        super.paintComponent(g);
-        
-        if (this.color == 1) {
-            g.setColor(new Color(221,192,127));
-        } else {
-            g.setColor(new Color(101,67,33));
+
+    public void capture(Piece piece) {
+        Piece existing = getOccupyingPiece();
+        if (existing != null) {
+            if (existing.getColor() == 0) {
+                board.getBlackPieces().remove(existing);
+            } else if (existing.getColor() == 1) {
+                board.getWhitePieces().remove(existing);
+            }
         }
-        
-        g.fillRect(this.getX(), this.getY(), this.getWidth(), this.getHeight());
-        
-        if(occupyingPiece != null && dispPiece) {
+        this.occupyingPiece = piece;
+    }
+
+    @Override
+    protected void paintComponent(Graphics g) {
+        super.paintComponent(g);
+
+        Color tileColor = (color == 1) ? new Color(221, 192, 127) : new Color(101, 67, 33);
+        g.setColor(tileColor);
+        g.fillRect(0, 0, getWidth(), getHeight());
+
+        if (occupyingPiece != null && displayPiece) {
             occupyingPiece.draw(g);
         }
     }
-    
+
     @Override
     public int hashCode() {
-        int prime = 31;
-        int result = 1;
-        result = prime * result + xNum;
-        result = prime * result + yNum;
-        return result;
+        return 31 * xNum + yNum;
     }
-    
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) return true;
+        if (!(obj instanceof Square square)) return false;
+        return this.xNum == square.xNum && this.yNum == square.yNum;
+    }
 }
