+++ src/View/GameWindow.java
@@ -1,3 +1,7 @@
+package View;
+
+import Model.Board;
+import Model.Clock;
 
 import java.awt.BorderLayout;
 import java.awt.GridLayout;
